requirements.txt
arrow
ipython
joblib
matplotlib
matplotlib-inline
numba
numpy
openpyxl
pillow
plotly
pyarrow
pygwalker
PyJWT
PyWavelets
PyYAML
scikit-learn
scipy
seaborn
shap
shapely
SQLAlchemy
sqlglot
sqlparse
statsmodels
streamlit
streamlit-aggrid
streamlit-authenticator
streamlit-on-Hover-tabs
streamlit-option-menu
toml
tqdm
xgboost
